[AN] refactor: calculateProgress 함수 위치 변경 #797
*   [AN] refactor: calculateProgress 함수를 HearitProgressBar.kt에서 Extensions.kt로 이동하여 공통화
